Backend/forecast/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.0
numpy==1.24.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
scikit-learn==1.3.0
xgboost==2.0.0
holidays==0.31
statsmodels==0.14.0
prophet==1.1.5
python-dotenv==1.0.0